# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Task",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "listing of the products",
                "operationId": "7e8cc1110886e25779f8f8235e5806cc",
                "requestBody": {
                    "description": "listing of the products",
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "json",
                                            "example": {
                                                "message": "success",
                                                "data": [
                                                    {
                                                        "id": 1,
                                                        "name": "product test 2",
                                                        "price": "2.00",
                                                        "created_at": "2023-06-24 13:07:28",
                                                        "updated_at": "2023-06-24 13:13:46"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid service id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "Invalid data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "something went wrong please try again later.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/products/store": {
            "post": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "store product",
                "operationId": "1cba87216678877e0aee200437e0ecce",
                "requestBody": {
                    "description": "store product",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "json",
                                            "example": {
                                                "message": "success",
                                                "data": {
                                                    "id": 5,
                                                    "name": "asd",
                                                    "price": "12.00",
                                                    "created_at": "2023-06-19T20:06:34.000000Z",
                                                    "updated_at": "2023-06-19T20:40:25.000000Z"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "Invalid Data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "something went wrong please try again later.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Get product",
                "operationId": "eee33694b51d083c745b05400887e2e1",
                "requestBody": {
                    "description": "store product",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "json",
                                            "example": {
                                                "message": "success",
                                                "data": {
                                                    "id": 5,
                                                    "name": "asd",
                                                    "price": "12.00",
                                                    "created_at": "2023-06-19T20:06:34.000000Z",
                                                    "updated_at": "2023-06-19T20:40:25.000000Z"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "Invalid Data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "something went wrong please try again later.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Delete Product",
                "operationId": "7f6fa0ad9c43e7b8a427d3c941a4c7ff",
                "requestBody": {
                    "description": "Delete Product",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "json",
                                            "example": {
                                                "message": "success",
                                                "data": []
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "Invalid Data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "something went wrong please try again later.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/product/update/{id}": {
            "put": {
                "tags": [
                    "Product Controller"
                ],
                "summary": "Update product",
                "operationId": "cc2d5cabfe02be418d35c05918268a1c",
                "requestBody": {
                    "description": "Update product",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "json",
                                            "example": {
                                                "message": "success",
                                                "data": {
                                                    "id": 5,
                                                    "name": "asd",
                                                    "price": "12.00",
                                                    "created_at": "2023-06-19T20:06:34.000000Z",
                                                    "updated_at": "2023-06-19T20:40:25.000000Z"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "Invalid Data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "something went wrong please try again later.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "example": "something went wrong please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}